+++ NEWS.md
@@ -6,6 +6,7 @@
 * Stricter Micronesia regular expressions. Thanks to @mattkerlogue, @stefgehrig, @cjyetman, @NilsEnevoldsen for discussion in Issue #354.
 * `TÜRKİYE` (dotted capital İ) now matches Turkey. Thanks to @cjyetman and @NilsEnevoldsen for report #347.
 * Update many dictionaries to the latest versions from publishing organizations. Thanks to @NilsEnevoldsen.
+* Improve Ireland regular expressions and add tests. Issue #313.
 
 ## countrycode 1.6.1